adding login via api and working-documents api

+++ observer/api.py
@@ -1,11 +1,37 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-from django.contrib.admin.views.decorators import staff_member_required
+import json
+
+from django.contrib.auth import REDIRECT_FIELD_NAME, login, logout, authenticate
+from django.contrib.auth.decorators import user_passes_test
 from django.db.models import Count, Q
+from django.http import HttpResponse
 
 from glue import Epoxy, API_EXCEPTION_AUTH, API_EXCEPTION_FORMERRORS, API_EXCEPTION_DOESNOTEXIST, API_EXCEPTION_HTTPERROR
-from observer.forms import DeviceForm, FullDocumentForm
-from observer.models import Device, Document
+from observer.forms import DeviceForm, FullDocumentForm, LoginForm
+from observer.models import Device, Document, WorkingDocument
+
+# 
+# decorator staff_member_required_for_POST
+# ===
+# Decorator for views that checks that the user is logged in and is a staff  member
+# only if is sending POST requests.
+#
+def staff_member_required_for_POST(view_func, redirect_field_name=REDIRECT_FIELD_NAME, login_url='observer_access_denied'):
+    """
+    
+    member, displaying the login page if necessary.
+    """
+    def decorator(*args, **kwargs):
+      if args[0].method == 'POST':
+        return user_passes_test(
+          lambda u: u.is_active and u.is_staff,
+          login_url=login_url,
+          redirect_field_name=redirect_field_name
+        )(view_func)
+
+      return view_func(*args, **kwargs)
+    return decorator
 
 
 # 
@@ -29,6 +55,7 @@ def is_number(s):
 #  @param django.db.models.Q instance
 #  @return <observer.models.Dpcument> or raise exception
 # 
+@staff_member_required_for_POST
 def get_available_document(request, q):
   '''
   @param request
@@ -55,6 +82,7 @@ def get_available_document(request, q):
 #  @param request
 #  @return django.model.Queryset
 # 
+@staff_member_required_for_POST
 def get_available_documents(request):
   
   if request.user.is_staff:
@@ -67,6 +95,27 @@ def get_available_documents(request):
 
 
 # 
+#  helper get_available_working_documents
+#  ===
+#
+#  Return a queryset according to user auth level and document status
+#  
+#  @param request
+#  @return django.model.Queryset
+#
+@staff_member_required_for_POST
+def get_available_working_documents(req):
+  
+  if req.user.is_staff:
+    queryset = WorkingDocument.objects.filter().distinct()
+  elif req.user.is_authenticated():
+    queryset = WorkingDocument.objects.filter(Q(status=WorkingDocument.PUBLIC) | Q(owner=req.user) | Q(authors=req.user)).distinct()
+  else:
+    queryset = WorkingDocument.objects.filter(status=WorkingDocument.PUBLIC).distinct()
+  return queryset
+
+
+# 
 #  api:/
 #  main entrance
 #
@@ -75,10 +124,51 @@ def index(req):
 
 
 # 
+#  api:/
+#  main entrance
+#
+def access_denied(req):
+  return Epoxy(req).throw_error(error='unauthorized', code=API_EXCEPTION_AUTH).json()
+
+
+# 
+#  api:/login
+#  main entrance
+#
+def login_view(req):
+  res = Epoxy(req)
+
+  if req.user.is_authenticated():
+    return res.json()
+
+  form = LoginForm(req.POST)
+  
+  #if not req.POST.get('remember_me', None):
+  #  req.session.set_expiry(0)
+
+  if form.is_valid():
+    user = authenticate(username=form.cleaned_data['username'], password=form.cleaned_data['password'])
+    if user is not None:
+      if user.is_active:
+        login(req, user)
+        # @todo: Redirect to next page
+        res.item(user)
+        return res.json()
+      else:
+        return res.throw_error(error='account disabled', code=API_EXCEPTION_AUTH).json()
+    else:
+      return res.throw_error(error='invalid credentials', code=API_EXCEPTION_AUTH).json()
+      # Return a 'disabled account' error message
+  else:
+    return res.throw_error(error=form.errors, code=API_EXCEPTION_AUTH).json()
+
+
+# 
 #  api:/document
 #  retrieve the requested documents and handle POST requests.
 #  At present, only staff can add documents.
 #
+@staff_member_required_for_POST
 def documents(req):
   res = Epoxy(req)
   if res.is_POST(): # staff only can add document via api
@@ -106,6 +196,7 @@ def documents(req):
 #  retrieve the requested document by numeric ID or by its SLUG field. and handle POST requests.
 #  At present, only staff can add documents.
 #
+@staff_member_required_for_POST
 def document(req, pk):
   res = Epoxy(request)
 
@@ -122,20 +213,73 @@ def document(req, pk):
   return res.item(doc, deep=True).json()
 
 
+# 
+#  api:/working-document
+#  retrieve the requested documents and handle POST requests.
+#  At present, only staff can add documents.
+#
+@staff_member_required_for_POST
+def workingDocuments(req):
+  res = Epoxy(req)
+  if req.user.is_staff and res.is_POST(): # staff only can add document via api
+    if not req.user.is_staff:
+      return res.throw_error(error='', code=API_EXCEPTION_AUTH).json()
+
+    form = FullDocumentForm(res.data)
+    if not form.is_valid():
+      return res.throw_error(error=form.errors, code=API_EXCEPTION_FORMERRORS).json()
+
+    doc = form.save(commit=False)
+    doc.owner = req.user
+
+    doc.save()
+    return res.item(doc).json()
+
+  queryset = get_available_working_documents(req)
+  res.queryset(queryset)
+  res.meta('query', '%s' % res._queryset.query)
+  return res.json()
+
+
+# 
+#  api:/working-document/(?P<pk>[:a-zA-Z\.\-\d]+)
+#  retrieve the requested working-document by numeric ID or by its SLUG field.
+#  At present, only staff can add new working-documents.
+#
+@staff_member_required_for_POST
+def workingDocument(req, pk):
+  res = Epoxy(request)
+
+  try:
+    doc = get_available_working_document(req, Q(pk=pk) if is_number(pk) else Q(slug=pk))
+  except Document.DoesNotExist, e:
+    return res.throw_error(error='%s' % e, code=API_EXCEPTION_DOESNOTEXIST).json()
+  
+  if epoxy.is_POST():
+    is_valid, doc = edit_object(instance=doc, Form=FullDocumentForm, request=req, epoxy=res)
+    if is_valid:
+      doc.save()
+
+  return res.item(doc, deep=True).json()
+
+
 #
 # api:/device
 # Get or create new device by using forms.DeviceForm
-# accessible by staff only
-@staff_member_required
+# A device is a connection between a Document and a Working Document
+# Post requests are accessible by staff only (create a new device)
+@staff_member_required_for_POST
 def devices(req):
   res = Epoxy(req)
   
   if res.is_POST():
+    if not req.user.is_staff:
+      return res.throw_error(error='', code=API_EXCEPTION_AUTH).json()
+
     form = DeviceForm(res.data)
-    if form.is_valid():
+    if is_valid:
       dev = form.save()
       res.item(dev)
-      return res.json()
     else:
       return res.throw_error(error=form.errors, code=API_EXCEPTION_FORMERRORS).json()
   
@@ -144,10 +288,10 @@ def devices(req):
 
 
 #
-# api:/device
-# Get or create new device by using forms.DeviceForm
-# accessible by staff only
-@staff_member_required
+# api:/device/(?P<pk>[:a-zA-Z\.\-\d]+)
+# Get,modify or delete a single device connection by using forms.DeviceForm
+# post requests are accessible by staff only
+@staff_member_required_for_POST
 def device(req, pk):
   res = Epoxy(req)
   
@@ -156,8 +300,12 @@ def device(req, pk):
   except Device.DoesNotExist, e:
     return res.throw_error(error='%s' % e, code=API_EXCEPTION_DOESNOTEXIST).json()
 
-  if res.is_DELETE():
+  if req.user.is_staff and res.is_DELETE():
     dev.delete()
+  if epoxy.is_POST():
+    is_valid, doc = edit_object(instance=doc, Form=FullDocumentForm, request=req, epoxy=res)
+    if is_valid:
+      doc.save()
   
   return res.item(doc, deep=True).json()
